fix(loading): show loading indicator when waiting for result
