+++ examples/query-playground/package.json
@@ -17,6 +17,7 @@
         "@dhis2/app-runtime": "*",
         "@dhis2/ui": "^5.5.3",
         "brace": "^0.11.1",
+        "classnames": "^2.2.6",
         "prop-types": "^15.7.2",
         "react": "^16.9.0",
         "react-ace": "^7.0.4",
