--- examples/query-playground/src/App.module.css
@@ -1,39 +1,0 @@
-.container {
-    position: absolute;
-    top: 48px;
-    bottom: 0px;
-    left: 0px;
-    right: 0px;
-    display: flex;
-    flex-direction: row;
-    align-items: flex-start;
-    justify-content: stretch;
-    font-size: 1rem;
-}
-
-.inner {
-    display: flex;
-    flex-direction: row;
-    align-items: stretch;
-
-    width: 100%;
-    height: 100%;
-}
-
-.editor {
-    width: 50%;
-    display: flex;
-    flex-direction: column;
-}
-
-.tabs {
-    flex-shrink: 1;
-}
-
-.currentTab {
-    flex-grow: 1;
-}
-
-.results {
-    width: 50%;
-}
