+++ examples/query-playground/src/components/GlobalLoading.module.css
@@ -1,4 +1,9 @@
 .container {
+    height: 100%;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
     position: absolute;
     top: 0;
     left: 0;
